requirements.txt
chainlitchainlit==1.1.304
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-groq==0.1.5
langchain-experimental==0.0.61
PyPDF2==3.0.1
chromadb==0.5.3
groq==0.9.0
ollama==0.2.1
pypandoc==1.13
pdfkit==1.0.0
docx2pdf==0.1.8
paddlepaddle==2.6.1
paddleocr==2.7.3
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.5
Faker==25.9.1
# python 3.10.0